test: split the reference resolvers out of the discovery suite
The resolvers decide which of a team's values a write will take, which is a
different subject from what the discovery endpoints serve -- and the tests for
them had landed in the middle of an 819-line file. Moved to their own file,
where the promise they keep can be stated as the module docstring: an id
/pipeline/options/ offered is an id a write takes, and nothing else is.
